Spring update

+++ springSecurityEazyBank/3eazyBankCorsCSRF/src/main/java/com/saha/amit/repository/AccountsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AccountsRepository extends CrudRepository<Accounts, Long> {
-	
-	Accounts findByCustomerId(int customerId);
+
+    Accounts findByCustomerId(long customerId);
 
 }
