+++ springSecurityEazyBank/3eazyBankCorsCSRF/src/main/java/com/saha/amit/repository/CardsRepository.java
@@ -1,14 +1,16 @@
 package com.saha.amit.repository;
 
-import com.saha.amit.model.Cards;
+import java.util.List;
+
+import com.saha.amit.model.Loans;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.saha.amit.model.Cards;
 
 @Repository
 public interface CardsRepository extends CrudRepository<Cards, Long> {
 	
-	List<Cards> findByCustomerId(int customerId);
+	List<Cards> findByCustomerId(long customerId);
 
 }
